requirements.txt
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.10